refactor(debugging): split profiler and test files into their respective folders
