+++ include/lvgl/debugging/profiler/lv_profiler_builtin.h
@@ -14,11 +14,11 @@ extern "C" {
  *      INCLUDES
  *********************/
 
-#include "../config/lv_conf_internal.h"
+#include "../../config/lv_conf_internal.h"
 
 #if LV_USE_PROFILER && LV_USE_PROFILER_BUILTIN
 
-#include "../lv_types.h"
+#include "../../lv_types.h"
 
 /*********************
  *      DEFINES
