+++ include/lvgl/debugging/test/lv_test_fs.h
@@ -14,10 +14,10 @@ extern "C" {
  *      INCLUDES
  *********************/
 
-#include "../config/lv_conf_internal.h"
+#include "../../config/lv_conf_internal.h"
 #if LV_USE_TEST
 
-#include "../lv_types.h"
+#include "../../lv_types.h"
 
 /*********************
  *      DEFINES
