+++ include/lvgl/debugging/test/lv_test_helpers.h
@@ -14,11 +14,11 @@ extern "C" {
  *      INCLUDES
  *********************/
 
-#include "../config/lv_conf_internal.h"
+#include "../../config/lv_conf_internal.h"
 #if LV_USE_TEST
 
-#include "../lv_types.h"
-#include "../stdlib/lv_mem.h"
+#include "../../lv_types.h"
+#include "../../stdlib/lv_mem.h"
 
 /*********************
  *      DEFINES
