+++ include/lvgl/debugging/test/lv_test_indev_gesture.h
@@ -13,11 +13,11 @@ extern "C" {
 /*********************
  *      INCLUDES
  *********************/
-#include "../config/lv_conf_internal.h"
+#include "../../config/lv_conf_internal.h"
 #if LV_USE_TEST && LV_USE_GESTURE_RECOGNITION
 
-#include "../lv_types.h"
-#include "../indev/lv_indev.h"
+#include "../../lv_types.h"
+#include "../../indev/lv_indev.h"
 
 /*********************
  *      DEFINES
