+++ include/lvgl/debugging/test/lv_test_screenshot_compare.h
@@ -13,7 +13,7 @@ extern "C" {
 /*********************
  *      INCLUDES
  *********************/
-#include "../config/lv_conf_internal.h"
+#include "../../config/lv_conf_internal.h"
 #if LV_USE_TEST && defined(LV_USE_TEST_SCREENSHOT_COMPARE) && LV_USE_TEST_SCREENSHOT_COMPARE
 
 /*********************
