+++ include/lvgl/lvgl.h
@@ -51,16 +51,16 @@
 #include "debugging/lv_assert.h"
 #include "debugging/lv_check_arg.h"
 #include "debugging/lv_monkey.h"
-#include "debugging/lv_profiler.h"
-#include "debugging/lv_profiler_builtin.h"
 #include "debugging/lv_sysmon.h"
-#include "debugging/lv_test.h"
-#include "debugging/lv_test_display.h"
-#include "debugging/lv_test_fs.h"
-#include "debugging/lv_test_helpers.h"
-#include "debugging/lv_test_indev.h"
-#include "debugging/lv_test_indev_gesture.h"
-#include "debugging/lv_test_screenshot_compare.h"
+#include "debugging/profiler/lv_profiler.h"
+#include "debugging/profiler/lv_profiler_builtin.h"
+#include "debugging/test/lv_test.h"
+#include "debugging/test/lv_test_display.h"
+#include "debugging/test/lv_test_fs.h"
+#include "debugging/test/lv_test_helpers.h"
+#include "debugging/test/lv_test_indev.h"
+#include "debugging/test/lv_test_indev_gesture.h"
+#include "debugging/test/lv_test_screenshot_compare.h"
 #include "display/lv_display.h"
 #include "draw/lv_color.h"
 #include "draw/lv_color_op.h"
